Interface (#74)
* Add prelude #63

* Make a step plot

* Add a new interface, Session

* Check dimensionality

* #19 Support a new keyword, proceed

* #19 Enable to extend reserved keywords

* Implement ensemble

* session.Result is picklable

* Use run_ensemble

* Move ecell4.util.viz -> ecell4.plotting

* ecell4.util.viz will be deprecated

* get_range_of_world and get_range_of_trajectories are moved into _core

* Copy deep

* Add plotting._plotly

* Add a datasource for STRING

* Add idmapping using UniProt
