+++ ecell4/util/__init__.py
@@ -2,11 +2,12 @@ import collections
 
 from .decorator import reaction_rules, species_attributes, get_model, reset_model
 from . import viz
-from .simulation import run_simulation, ensemble_simulations, load_world, number_observer
+from .simulation import run_simulation, ensemble_simulations, number_observer
 from . import ports
 from .show import show
+from .session import Session, load_world
 
 __all__ = [
     'run_simulation', 'ensemble_simulations', 'load_world', 'number_observer',
     'reaction_rules', 'species_attributes', 'get_model', 'show',
-    'viz', 'ports']
+    'viz', 'ports', 'Session']
