+++ setup.py
@@ -12,7 +12,7 @@ LONG_DESCRIPTION = open(os.path.join(".", "README.md")).read()
 setup(
     name='ecell4',
     version='1.1.2',
-    packages=['ecell4', 'ecell4.util', 'ecell4.extra', 'ecell4.datasource', 'ecell4.mca'],
+    packages=['ecell4', 'ecell4.util', 'ecell4.extra', 'ecell4.datasource', 'ecell4.mca', 'ecell4.plotting'],
     package_data = {"ecell4.util": [
         "templates/init_ipynb.js", "templates/init_cyjs.js", "templates/template.html",
         "templates/*.tmpl", "templates/ecelllogo/*.png"]},
